requirements.txt
django==3.1
pillow==7.2.0
psycopg2-binary==2.8.6
django-environ==0.4.5
django-crispy-forms==1.11.1
django-allauth==0.44.0
django-sequences==2.6